feat(cli): move cmd_* functions into pyquotex.cli.commands
Distributes the 27 cmd_* CLI handlers from app.py into seven per-domain
modules under pyquotex/cli/commands/. The COMMAND_REGISTRY dict maps
argparse subcommand names to handlers. app.py is left non-functional
until Phase 4.3 wires the new pyquotex.cli.__main__ entry point and
Phase 4.4 reduces app.py to a thin shim.
